backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
pydantic
python-dotenv
bcrypt
requests
sqlalchemy
aiofiles
MetaTrader5
python-socketio
asyncio-mqtt
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]
redis
celery 